+++ include/PullStateManager.h
@@ -20,6 +20,7 @@ public:
 
   // External triggers
   static void triggerEmergencyStop();
+  static void triggerAlarmDqStop();
   static void detectPullStart(float currentSpeed);  // Can be called periodically to detect transition to PULLING
   static void updateUIForState(PullState state);
   static void enterState(PullState newState);